chore: fix code typo and move project details components under projects/[id]/_components
