+++ src/app/(main)/projects/[id]/_components/contract-details.stories.tsx
